fix(protocol/wire): decode_hardlink_idx return type i32 → u32
The hardlink group index is a non-negative offset into the file list.
The previous Option<i32> return type forced callers to cast back to
usize; for indices above 0x7FFF_FFFF the reinterpretation is negative,
which either panics on the subsequent flist[] access or silently links
the wrong file without any error.

Change the return type to Option<u32> and apply a matching 'as u32'
cast on the read_varint result.  The encoder (encode_hardlink_idx)
already accepts u32 and the wire value is always non-negative.

Upstream rsync uses unsigned int for hlink_flist indices.

+++ crates/protocol/src/wire/file_entry_decode.rs
@@ -791,14 +791,16 @@ pub fn decode_symlink_target<R: Read>(reader: &mut R, protocol_version: u8) -> i
 ///
 /// Only decode when `XMIT_HLINKED` is set but `XMIT_HLINK_FIRST` is NOT set.
 /// The first occurrence of a hardlink group (leader) doesn't have an index.
-pub fn decode_hardlink_idx<R: Read>(reader: &mut R, flags: u32) -> io::Result<Option<i32>> {
+pub fn decode_hardlink_idx<R: Read>(reader: &mut R, flags: u32) -> io::Result<Option<u32>> {
     if flags & ((XMIT_HLINKED as u32) << 8) != 0 {
         if flags & ((XMIT_HLINK_FIRST as u32) << 8) != 0 {
             // First occurrence (leader) - no index follows
             Ok(None)
         } else {
-            // Follower - read index
-            Ok(Some(read_varint(reader)?))
+            // Follower - read index.
+            // Cast i32 bits to u32 to preserve the full index space;
+            // upstream C uses unsigned int for hlink_flist indices.
+            Ok(Some(read_varint(reader)? as u32))
         }
     } else {
         Ok(None)
